getting the balance factor of a binary tree

+++ 14-binary_tree_balance.c
@@ -15,7 +15,7 @@
 
 int binary_tree_balance(const binary_tree_t *tree)
 {
-	int left_height = 0, right_height = 0;
+	size_t left_height = 0, right_height = 0;
 
 	if (tree == NULL)
 		return (0);
